+++ share/objects/slate/elder.lua
@@ -24,3 +24,23 @@ do
 	-- Duration - Normal
 	debuff.adur.value = duration
 end
+
+-- # Elder Information
+do
+	local info = objects ['A061']
+
+	-- ## Text
+
+	-- Tooltip - Normal - Extended
+	info.aub1.value = {
+		type = 'string',
+		value = 'Chance to cast per attack:|n'
+			.. '- Frost Nova|n'
+			.. '- Thunder Clap|n'
+			.. '- Fan of Knives|n'
+			.. '- Armor Reduction|n'
+			.. '- Shockwave|n'
+			.. '- Forked Lightning|n'
+			.. '- Poison Shot'
+	}
+end