+++ docs/changelog.md
@@ -18,6 +18,7 @@ full list of changes to the project, please consult the commit log.
       rather than giving the target Slate directly.
 - Elder Slate:
     - Armor debuff duration no longer increases with levels.
+    - No longer rewards gold.
 
 ## [1.4.3] - 2018-04-13