+++ docs/changelog.md
@@ -19,6 +19,8 @@ full list of changes to the project, please consult the commit log.
 - Elder Slate:
     - Armor debuff duration no longer increases with levels.
     - No longer rewards gold.
+- Spell Slate:
+    - No longer reward gold.
 
 ## [1.4.3] - 2018-04-13