+++ docs/changelog.md
@@ -5,6 +5,10 @@ full list of changes to the project, please consult the commit log.
 
 ## [Unreleased]
 
+## Fixed
+- Dead gems are no longer replaced by rocks upon using action buttons (e.g.
+  Keep or Combine).
+
 ### Changed
 - Combine has been reworked:
     - Existing functionality has been split into distinct Combine (2x) and